move cf specific code to cloudformation.go
